вынес функции для работы со списками t_list в t_list_utils.c
